run dash from notebook
