+++ geoapps/scatter_plot/application.py
@@ -4,45 +4,30 @@
 #
 #  geoapps is distributed under the terms and conditions of the MIT License
 #  (see LICENSE file at the root of this source code package).
-
+import base64
+import io
 import os
+import webbrowser
+from os import environ
 
+import dash
 import numpy as np
 import plotly.express as px
 import plotly.graph_objects as go
+from dash import dcc, html
+from dash.dependencies import Input, Output
+from flask import Flask
+from geoh5py.objects.object_base import ObjectBase
 from geoh5py.ui_json import InputFile
-from ipywidgets import (
-    Checkbox,
-    Dropdown,
-    FloatText,
-    HBox,
-    IntSlider,
-    Label,
-    Layout,
-    ToggleButton,
-    VBox,
-)
+from geoh5py.workspace import Workspace
 
-from geoapps.base.selection import ObjectDataSelection
 from geoapps.scatter_plot.constants import app_initializer
 from geoapps.scatter_plot.driver import ScatterPlotDriver
 from geoapps.scatter_plot.params import ScatterPlotParams
 
 
-class ScatterPlots(ObjectDataSelection):
-    """
-    Application for 2D and 3D crossplots of data using symlog scaling
-    """
-
+class ScatterPlots:
     _param_class = ScatterPlotParams
-    _select_multiple = True
-    _add_groups = False
-    _downsampling = None
-    _color = None
-    _x = None
-    _y = None
-    _z = None
-    _size = None
 
     def __init__(self, ui_json=None, **kwargs):
         app_initializer.update(kwargs)
@@ -51,500 +36,775 @@ class ScatterPlots(ObjectDataSelection):
         else:
             self.params = self._param_class(**app_initializer)
 
-        self.defaults = {}
-        for key, value in self.params.to_dict().items():
-            self.defaults[key] = value
-
-        def channel_bounds_setter(caller):
-            self.set_channel_bounds(caller["owner"].name)
-
-        self.x.observe(channel_bounds_setter, names="value")
-        self.x.name = "x"
-        self._x_log = Checkbox(
-            description="Log10",
-            value=False,
-            indent=False,
-        )
-        self._x_thresh = FloatText(
-            description="Threshold",
-            value=1e-1,
-        )
-        self._x_min = FloatText(
-            description="Min",
-        )
-        self._x_max = FloatText(
-            description="Max",
-        )
-        self._x_panel = VBox(
-            [
-                HBox([self._x]),
-                HBox([self._x_log, self._x_thresh]),
-                HBox([self._x_min, self._x_max]),
-            ]
-        )
-        self.y.observe(channel_bounds_setter, names="value")
-        self.y.name = "y"
-        self._y_log = Checkbox(description="Log10", value=False, indent=False)
-        self._y_thresh = FloatText(
-            description="Threshold",
-            value=1e-1,
-        )
-        self._y_min = FloatText(
-            description="Min",
-        )
-        self._y_max = FloatText(
-            description="Max",
-        )
-        self._y_panel = VBox(
-            [
-                HBox([self._y]),
-                HBox([self._y_log, self._y_thresh]),
-                HBox([self._y_min, self._y_max]),
-            ]
-        )
-        self.z.observe(channel_bounds_setter, names="value")
-        self.z.name = "z"
-        self._z_log = Checkbox(
-            description="Log10",
-            value=False,
-            indent=False,
-        )
-        self._z_thresh = FloatText(
-            description="Threshold",
-            value=1e-1,
-        )
-        self._z_min = FloatText(
-            description="Min",
-        )
-        self._z_max = FloatText(
-            description="Max",
-        )
-        self._z_panel = VBox(
-            [
-                HBox([self._z]),
-                HBox([self._z_log, self._z_thresh]),
-                HBox([self._z_min, self._z_max]),
-            ]
-        )
-        self.color.observe(channel_bounds_setter, names="value")
-        self.color.name = "color"
-        self._color_log = Checkbox(
-            description="Log10",
-            value=False,
-            indent=False,
-        )
-        self._color_thresh = FloatText(
-            description="Threshold",
-            value=1e-1,
-        )
-        self._color_maps = Dropdown(
-            description="Colormaps",
-            options=px.colors.named_colorscales(),
-            value="viridis",
-        )
-        self._color_min = FloatText(
-            description="Min",
-        )
-        self._color_max = FloatText(
-            description="Max",
-        )
-        self._color_panel = VBox(
-            [
-                HBox([self._color]),
-                self._color_maps,
-                HBox([self._color_log, self._color_thresh]),
-                HBox([self._color_min, self._color_max]),
-            ]
-        )
-        self.size.observe(channel_bounds_setter, names="value")
-        self.size.name = "size"
-        self._size_log = Checkbox(
-            description="Log10",
-            value=False,
-            indent=False,
-        )
-        self._size_thresh = FloatText(
-            description="Threshold",
-            value=1e-1,
-        )
-        self._size_markers = IntSlider(
-            min=1, max=100, value=20, description="Marker size", continuous_update=False
-        )
-        self._size_min = FloatText(
-            description="Min",
-        )
-        self._size_max = FloatText(
-            description="Max",
-        )
-
-        self._size_panel = VBox(
-            [
-                HBox([self._size]),
-                self._size_markers,
-                HBox([self._size_log, self._size_thresh]),
-                HBox([self._size_min, self._size_max]),
-            ]
-        )
-        self._refresh = ToggleButton(description="Refresh Plot", value=False)
-
-        # Wrap all axis panels into dropdown
-        def axes_pannels_trigger(_):
-            self.axes_pannels_trigger()
-
-        self.panels = {
-            "X-axis": self._x_panel,
-            "Y-axis": self._y_panel,
-            "Z-axis": self._z_panel,
-            "Color": self._color_panel,
-            "Size": self._size_panel,
-        }
-        self.axes_pannels = Dropdown(
-            options=["X-axis", "Y-axis", "Z-axis", "Color", "Size"],
-            layout=Layout(width="300px"),
-        )
-        self.axes_pannels.observe(axes_pannels_trigger, names="value")
-        self.axes_options = VBox([self.axes_pannels, self._x_panel])
         self.data_channels = {}
-        self.objects.observe(self.update_objects, names="value")
-        self.objects.observe(self.update_choices, names="value")
-        self.figure = go.FigureWidget()
 
-        self.downsampling.observe(self.plot_selection, names="value")
-        self.x.observe(self.plot_selection, names="value")
-        self.x_log.observe(self.plot_selection, names="value")
-        self.x_thresh.observe(self.plot_selection, names="value")
-        self.x_min.observe(self.plot_selection, names="value")
-        self.x_max.observe(self.plot_selection, names="value")
-        self.y.observe(self.plot_selection, names="value")
-        self.y_log.observe(self.plot_selection, names="value")
-        self.y_thresh.observe(self.plot_selection, names="value")
-        self.y_min.observe(self.plot_selection, names="value")
-        self.y_max.observe(self.plot_selection, names="value")
-        self.z.observe(self.plot_selection, names="value")
-        self.z_log.observe(self.plot_selection, names="value")
-        self.z_thresh.observe(self.plot_selection, names="value")
-        self.z_min.observe(self.plot_selection, names="value")
-        self.z_max.observe(self.plot_selection, names="value")
-        self.color.observe(self.plot_selection, names="value")
-        self.color_log.observe(self.plot_selection, names="value")
-        self.color_thresh.observe(self.plot_selection, names="value")
-        self.color_min.observe(self.plot_selection, names="value")
-        self.color_max.observe(self.plot_selection, names="value")
-        self.color_maps.observe(self.plot_selection, names="value")
-        self.size.observe(self.plot_selection, names="value")
-        self.size_log.observe(self.plot_selection, names="value")
-        self.size_thresh.observe(self.plot_selection, names="value")
-        self.size_min.observe(self.plot_selection, names="value")
-        self.size_max.observe(self.plot_selection, names="value")
-        self.size_markers.observe(self.plot_selection, names="value")
+        # Initial values for the dash components
+        self.defaults = {}
+        self.set_defaults()
 
-        self.trigger.on_click(self.trigger_click)
-        self.trigger.description = "Save HTML"
+        external_stylesheets = ["https://codepen.io/chriddyp/pen/bWLwgP.css"]
+        server = Flask(__name__)
+        self.app = dash.Dash(
+            server=server,
+            url_base_pathname=environ.get("JUPYTERHUB_SERVICE_PREFIX", "/"),
+            external_stylesheets=external_stylesheets,
+        )
 
-        super().__init__(**self.defaults)
+        # Set up the layout with the dash components
+        self.app.layout = html.Div(
+            [
+                html.Div(
+                    [
+                        dcc.Markdown(
+                            children="""
+                            ### Scatter Plots
 
-        self._x.value = self.params.x.name
-        self._y.value = self.params.y.name
-        self._z.value = self.params.z.name
-        self._color.value = self.params.color.name
-        self._size.value = self.params.size.name
+                            This application lets users visualize up to 5D of data pulled from any Geoscience ANALYST objects. The application uses the rich [Plotly](https://plotly.com/) graphical interface.
 
-    @property
-    def color(self):
-        """
-        Widget for the selection of color values
-        """
-        if getattr(self, "_color", None) is None:
-            self._color = Dropdown(description="Data:")
+                            New user? Visit the [**Getting Started**](https://geoapps.readthedocs.io/en/latest/content/installation.html) page.
 
-        return self._color
+                            [**Online Documentation**](https://geoapps.readthedocs.io/en/latest/content/applications/scatter.html)
+                            """
+                        ),
+                    ],
+                    style={
+                        "width": "100%",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    [
+                        dcc.Upload(
+                            id="upload",
+                            children=html.Button("Change Workspace"),
+                            style={"margin-bottom": "20px"},
+                        ),
+                        dcc.Dropdown(
+                            id="objects",
+                            options=self.defaults["object_options"],
+                            value=self.defaults["objects_name"],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Population Downsampling (%): ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Slider(
+                                    id="downsampling",
+                                    value=self.defaults["downsampling"],
+                                    min=1,
+                                    max=100,
+                                    step=1,
+                                    marks=None,
+                                    tooltip={
+                                        "placement": "bottom",
+                                        "always_visible": True,
+                                    },
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        dcc.Dropdown(
+                            id="axes_pannels",
+                            options=[
+                                {"label": "X-axis", "value": "x"},
+                                {"label": "Y-axis", "value": "y"},
+                                {"label": "Z-axis", "value": "z"},
+                                {"label": "Color", "value": "color"},
+                                {"label": "Size", "value": "size"},
+                            ],
+                            value="x",
+                            style={"margin-bottom": "20px"},
+                        ),
+                    ],
+                    style={"width": "40%", "display": "block", "vertical-align": "top"},
+                ),
+                html.Div(
+                    id="x_div",
+                    children=[
+                        dcc.Markdown(children="Data: "),
+                        dcc.Dropdown(
+                            id="x",
+                            options=self.defaults["data_options"],
+                            value=self.defaults["x_name"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Threshold: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="x_thresh",
+                                    type="number",
+                                    value=self.defaults["x_thresh"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Checklist(
+                                    id="x_log",
+                                    options=["Log10"],
+                                    value=self.defaults["x_log"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Min: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="x_min",
+                                    type="number",
+                                    value=self.defaults["x_min"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Markdown(
+                                    children="Max: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="x_max",
+                                    type="number",
+                                    value=self.defaults["x_max"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                        ),
+                    ],
+                    style={
+                        "display": "block",
+                        "width": "40%",
+                        "vertical-align": "top",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    id="y_div",
+                    children=[
+                        dcc.Markdown(children="Data: "),
+                        dcc.Dropdown(
+                            id="y",
+                            options=self.defaults["data_options"],
+                            value=self.defaults["y_name"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Threshold: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="y_thresh",
+                                    type="number",
+                                    value=self.defaults["y_thresh"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Checklist(
+                                    id="y_log",
+                                    options=["Log10"],
+                                    value=self.defaults["y_log"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Min: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="y_min",
+                                    type="number",
+                                    value=self.defaults["y_min"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Markdown(
+                                    children="Max: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="y_max",
+                                    type="number",
+                                    value=self.defaults["y_max"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                        ),
+                    ],
+                    style={
+                        "display": "none",
+                        "width": "40%",
+                        "vertical-align": "top",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    id="z_div",
+                    children=[
+                        dcc.Markdown(children="Data: "),
+                        dcc.Dropdown(
+                            id="z",
+                            options=self.defaults["data_options"],
+                            value=self.defaults["z_name"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Threshold: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="z_thresh",
+                                    type="number",
+                                    value=self.defaults["z_thresh"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Checklist(
+                                    id="z_log",
+                                    options=["Log10"],
+                                    value=self.defaults["z_log"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Min: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="z_min",
+                                    type="number",
+                                    value=self.defaults["z_min"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Markdown(
+                                    children="Max: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="z_max",
+                                    type="number",
+                                    value=self.defaults["z_max"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                        ),
+                    ],
+                    style={
+                        "display": "none",
+                        "width": "40%",
+                        "vertical-align": "top",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    id="color_div",
+                    children=[
+                        dcc.Markdown(children="Data: "),
+                        dcc.Dropdown(
+                            id="color",
+                            options=self.defaults["data_options"],
+                            value=self.defaults["color_name"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        dcc.Dropdown(
+                            id="color_maps",
+                            options=px.colors.named_colorscales(),
+                            value=self.defaults["color_maps"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Threshold: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="color_thresh",
+                                    type="number",
+                                    value=self.defaults["color_thresh"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Checklist(
+                                    id="color_log",
+                                    options=["Log10"],
+                                    value=self.defaults["color_log"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Min: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="color_min",
+                                    type="number",
+                                    value=self.defaults["color_min"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Markdown(
+                                    children="Max: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="color_max",
+                                    type="number",
+                                    value=self.defaults["color_max"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                        ),
+                    ],
+                    style={
+                        "display": "none",
+                        "width": "40%",
+                        "vertical-align": "top",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    id="size_div",
+                    children=[
+                        dcc.Markdown(children="Data: "),
+                        dcc.Dropdown(
+                            id="size",
+                            options=self.defaults["data_options"],
+                            value=self.defaults["size_name"],
+                            style={"width": "63.3%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Marker Size: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Slider(
+                                    id="size_markers",
+                                    value=self.defaults["size_markers"],
+                                    min=1,
+                                    max=100,
+                                    step=1,
+                                    marks=None,
+                                    tooltip={
+                                        "placement": "bottom",
+                                        "always_visible": True,
+                                    },
+                                ),
+                            ],
+                            style={"width": "40%", "margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Threshold: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="size_thresh",
+                                    type="number",
+                                    value=self.defaults["size_thresh"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Checklist(
+                                    id="size_log",
+                                    options=["Log10"],
+                                    value=self.defaults["size_log"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                            style={"margin-bottom": "20px"},
+                        ),
+                        html.Div(
+                            [
+                                dcc.Markdown(
+                                    children="Min: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="size_min",
+                                    type="number",
+                                    value=self.defaults["size_min"],
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "20px",
+                                    },
+                                ),
+                                dcc.Markdown(
+                                    children="Max: ",
+                                    style={
+                                        "display": "inline-block",
+                                        "margin-right": "5px",
+                                    },
+                                ),
+                                dcc.Input(
+                                    id="size_max",
+                                    type="number",
+                                    value=self.defaults["size_max"],
+                                    style={"display": "inline-block"},
+                                ),
+                            ],
+                        ),
+                    ],
+                    style={
+                        "display": "none",
+                        "width": "40%",
+                        "vertical-align": "top",
+                        "margin-bottom": "20px",
+                    },
+                ),
+                html.Div(
+                    [
+                        dcc.Graph(
+                            id="plot",
+                        ),
+                        html.A(
+                            html.Button("Download as HTML"),
+                            id="download",
+                            download="Crossplot.html",
+                        ),
+                    ],
+                    style={"width": "80%", "display": "block", "margin-bottom": "20px"},
+                ),
+                dcc.Markdown(
+                    children="""Need help? Contact us at support@mirageoscience.com"""
+                ),
+            ],
+            style={"width": "100%", "margin-left": "50px"},
+        )
 
-    @property
-    def color_log(self):
-        """
-        :obj:`ipywidgets.Checkbox`
-        """
-        return self._color_log
+        # Set up callbacks
+        self.app.callback(
+            Output(component_id="x_div", component_property="style"),
+            Output(component_id="y_div", component_property="style"),
+            Output(component_id="z_div", component_property="style"),
+            Output(component_id="color_div", component_property="style"),
+            Output(component_id="size_div", component_property="style"),
+            Input(component_id="axes_pannels", component_property="value"),
+        )(self.update_visibility)
+        self.app.callback(
+            Output(component_id="x_min", component_property="value"),
+            Output(component_id="x_max", component_property="value"),
+            Output(component_id="y_min", component_property="value"),
+            Output(component_id="y_max", component_property="value"),
+            Output(component_id="z_min", component_property="value"),
+            Output(component_id="z_max", component_property="value"),
+            Output(component_id="color_min", component_property="value"),
+            Output(component_id="color_max", component_property="value"),
+            Output(component_id="size_min", component_property="value"),
+            Output(component_id="size_max", component_property="value"),
+            Input(component_id="x", component_property="value"),
+            Input(component_id="y", component_property="value"),
+            Input(component_id="z", component_property="value"),
+            Input(component_id="color", component_property="value"),
+            Input(component_id="size", component_property="value"),
+        )(self.set_channel_bounds)
+        self.app.callback(
+            Output(component_id="plot", component_property="figure"),
+            Input(component_id="downsampling", component_property="value"),
+            Input(component_id="x", component_property="value"),
+            Input(component_id="x_log", component_property="value"),
+            Input(component_id="x_thresh", component_property="value"),
+            Input(component_id="x_min", component_property="value"),
+            Input(component_id="x_max", component_property="value"),
+            Input(component_id="y", component_property="value"),
+            Input(component_id="y_log", component_property="value"),
+            Input(component_id="y_thresh", component_property="value"),
+            Input(component_id="y_min", component_property="value"),
+            Input(component_id="y_max", component_property="value"),
+            Input(component_id="z", component_property="value"),
+            Input(component_id="z_log", component_property="value"),
+            Input(component_id="z_thresh", component_property="value"),
+            Input(component_id="z_min", component_property="value"),
+            Input(component_id="z_max", component_property="value"),
+            Input(component_id="color", component_property="value"),
+            Input(component_id="color_log", component_property="value"),
+            Input(component_id="color_thresh", component_property="value"),
+            Input(component_id="color_min", component_property="value"),
+            Input(component_id="color_max", component_property="value"),
+            Input(component_id="color_maps", component_property="value"),
+            Input(component_id="size", component_property="value"),
+            Input(component_id="size_log", component_property="value"),
+            Input(component_id="size_thresh", component_property="value"),
+            Input(component_id="size_min", component_property="value"),
+            Input(component_id="size_max", component_property="value"),
+            Input(component_id="size_markers", component_property="value"),
+        )(self.plot_selection)
+        self.app.callback(
+            Output(component_id="objects", component_property="options"),
+            Output(component_id="objects", component_property="value"),
+            Input(component_id="upload", component_property="contents"),
+            prevent_initial_call=True,
+        )(self.update_object_options)
+        self.app.callback(
+            Output(component_id="x", component_property="options"),
+            Output(component_id="y", component_property="options"),
+            Output(component_id="z", component_property="options"),
+            Output(component_id="color", component_property="options"),
+            Output(component_id="size", component_property="options"),
+            Output(component_id="x", component_property="value"),
+            Output(component_id="y", component_property="value"),
+            Output(component_id="z", component_property="value"),
+            Output(component_id="color", component_property="value"),
+            Output(component_id="size", component_property="value"),
+            Input(component_id="objects", component_property="value"),
+            prevent_initial_call=True,
+        )(self.update_data_options)
+        self.app.callback(
+            Output(component_id="download", component_property="href"),
+            Input(component_id="plot", component_property="figure"),
+        )(self.save_figure)
 
-    @property
-    def color_thresh(self):
-        """
-        :obj:`ipywidgets.FloatText`
-        """
-        return self._color_thresh
-
-    @property
-    def color_maps(self):
-        """
-        :obj:`ipywidgets.Dropdown`
-        """
-        return self._color_maps
-
-    @property
-    def color_min(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._color_min
-
-    @property
-    def color_max(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._color_max
-
-    @property
-    def downsampling(self) -> IntSlider:
-        """
-        Widget controlling the size of the population.
-        """
-        if getattr(self, "_downsampling", None) is None:
-            self._downsampling = IntSlider(
-                description="Population Downsampling (%):",
-                min=1,
-                max=100,
-                style={"description_width": "initial"},
-                continuous_update=False,
-                value=100,
+    def update_visibility(self, axis):
+        # Change the visibility of the dash components depending on the axis selected
+        if axis == "x":
+            return (
+                {"display": "block"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
             )
-        return self._downsampling
+        elif axis == "y":
+            return (
+                {"display": "none"},
+                {"display": "block"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
+            )
+        elif axis == "z":
+            return (
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "block"},
+                {"display": "none"},
+                {"display": "none"},
+            )
+        elif axis == "color":
+            return (
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "block"},
+                {"display": "none"},
+            )
+        elif axis == "size":
+            return (
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "none"},
+                {"display": "block"},
+            )
 
-    @property
-    def main(self):
-        if self._main is None:
-            self._main = VBox(
-                [
-                    self.project_panel,
-                    self.objects,
-                    VBox([Label("Downsampling"), self.downsampling]),
-                    self.axes_options,
-                    self.figure,
-                    self.trigger,
+    def set_defaults(self):
+        # Get initial values to initialize the dash components
+        for key, value in self.params.to_dict().items():
+            if key in ["x", "y", "z", "color", "size"]:
+                self.defaults[key + "_name"] = value.name
+            elif key == "objects":
+                self.defaults[key + "_name"] = value.name
+                data_options = ["None"]
+                data_options.extend([data.name for data in value.children])
+                if "Visual Parameters" in data_options:
+                    data_options.remove("Visual Parameters")
+                self.defaults["data_options"] = data_options
+            elif key in ["x_log", "y_log", "z_log", "color_log", "size_log"]:
+                if value is True:
+                    self.defaults[key] = ["Log10"]
+                else:
+                    self.defaults[key] = []
+            else:
+                self.defaults[key] = value
+
+            if key == "geoh5":
+                self.defaults["object_options"] = [
+                    {"label": obj.parent.name + "/" + obj.name, "value": obj.name}
+                    for obj in value.objects
                 ]
-            )
-
-        return self._main
-
-    @property
-    def refresh(self):
-        """"""
-        return self._refresh
-
-    @property
-    def size(self):
-        """
-        Widget for the selection of size scaling values
-        """
-        if getattr(self, "_size", None) is None:
-            self._size = Dropdown(description="Data:")
-        return self._size
-
-    @property
-    def size_log(self):
-        """
-        :obj:`ipywidgets.Checkbox`
-        """
-        return self._size_log
-
-    @property
-    def size_thresh(self):
-        """
-        :obj:`ipywidgets.FloatText`
-        """
-        return self._size_thresh
-
-    @property
-    def size_markers(self):
-        """
-        :obj:`ipywidgets.IntSlider`
-        """
-        return self._size_markers
-
-    @property
-    def size_min(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._size_min
-
-    @property
-    def size_max(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._size_max
-
-    @property
-    def x(self):
-        """
-        Widget for the selection of x-axis values
-        """
-        if getattr(self, "_x", None) is None:
-            self._x = Dropdown(description="Data:")
-        return self._x
-
-    @property
-    def x_log(self):
-        """
-        :obj:`ipywidgets.Checkbox`
-        """
-        return self._x_log
-
-    @property
-    def x_thresh(self):
-        """
-        :obj:`ipywidgets.FloatText`
-        """
-        return self._x_thresh
-
-    @property
-    def x_min(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._x_min
-
-    @property
-    def x_max(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._x_max
-
-    @property
-    def y(self):
-        """
-        Widget for the selection of y-axis values
-        """
-        if getattr(self, "_y", None) is None:
-            self._y = Dropdown(description="Data:")
-        return self._y
-
-    @property
-    def y_log(self):
-        """
-        :obj:`ipywidgets.Checkbox`
-        """
-        return self._y_log
-
-    @property
-    def y_thresh(self):
-        """
-        :obj:`ipywidgets.FloatText`
-        """
-        return self._y_thresh
-
-    @property
-    def y_min(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._y_min
-
-    @property
-    def y_max(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._y_max
-
-    @property
-    def z(self):
-        """
-        Widget for the selection of z-axis values
-        """
-        if getattr(self, "_z", None) is None:
-            self._z = Dropdown(description="Data:")
-        return self._z
-
-    @property
-    def z_log(self):
-        """
-        :obj:`ipywidgets.Checkbox`
-        """
-        return self._z_log
-
-    @property
-    def z_thresh(self):
-        """
-        :obj:`ipywidgets.FloatText`
-        """
-        return self._z_thresh
-
-    @property
-    def z_min(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._z_min
-
-    @property
-    def z_max(self):
-        """
-        :obj:`ipywidgets.Text`
-        """
-        return self._z_max
-
-    def axes_pannels_trigger(self):
-        self.axes_options.children = [
-            self.axes_pannels,
-            self.panels[self.axes_pannels.value],
-        ]
 
     def get_channel(self, channel):
-        obj, _ = self.get_selected_entities()
-
         if channel is None:
             return None
 
         if channel not in self.data_channels.keys():
             if channel == "None":
                 data = None
-            elif self.workspace.get_entity(channel):
-                data = self.workspace.get_entity(channel)[0]
+            elif self.params.geoh5.get_entity(channel):
+                data = self.params.geoh5.get_entity(channel)[0]
             else:
                 return
 
             self.data_channels[channel] = data
 
-    def set_channel_bounds(self, name):
+    def get_channel_bounds(self, channel):
         """
         Set the min and max values for the given axis channel
         """
-        self.refresh.value = False
-
-        channel = getattr(self, "_" + name).value
         self.get_channel(channel)
 
-        if channel in self.data_channels.keys():
-            if channel == "None":
-                cmin = getattr(self, "_" + name + "_min")
-                cmin.value = 0
-                cmax = getattr(self, "_" + name + "_max")
-                cmax.value = 0
-            else:
-                values = self.data_channels[channel].values
-                values = values[~np.isnan(values)]
-                cmin = getattr(self, "_" + name + "_min")
-                cmin.value = f"{np.min(values):.2e}"
-                cmax = getattr(self, "_" + name + "_max")
-                cmax.value = f"{np.max(values):.2e}"
+        cmin, cmax = 0, 0
+        if (channel in self.data_channels.keys()) & (channel != "None"):
+            values = self.data_channels[channel].values
+            values = values[~np.isnan(values)]
+            cmin = f"{np.min(values):.2e}"
+            cmax = f"{np.max(values):.2e}"
 
-        self.refresh.value = True
+        return cmin, cmax
 
-    def plot_selection(self, _):
-        if not self.refresh.value:
-            return None
+    def set_channel_bounds(self, x, y, z, color, size):
+        x_min, x_max = self.get_channel_bounds(x)
+        y_min, y_max = self.get_channel_bounds(y)
+        z_min, z_max = self.get_channel_bounds(z)
+        color_min, color_max = self.get_channel_bounds(color)
+        size_min, size_max = self.get_channel_bounds(size)
 
+        return (
+            x_min,
+            x_max,
+            y_min,
+            y_max,
+            z_min,
+            z_max,
+            color_min,
+            color_max,
+            size_min,
+            size_max,
+        )
+
+    def plot_selection(
+        self,
+        downsampling,
+        x,
+        x_log,
+        x_thresh,
+        x_min,
+        x_max,
+        y,
+        y_log,
+        y_thresh,
+        y_min,
+        y_max,
+        z,
+        z_log,
+        z_thresh,
+        z_min,
+        z_max,
+        color,
+        color_log,
+        color_thresh,
+        color_min,
+        color_max,
+        color_maps,
+        size,
+        size_log,
+        size_thresh,
+        size_min,
+        size_max,
+        size_markers,
+    ):
         new_params_dict = {}
         for key, value in self.params.to_dict().items():
-            param = getattr(self, key, None)
+
+            if key in locals():
+                param = locals()[key]
+            else:
+                param = None
 
             if param is None:
                 new_params_dict[key] = value
-            elif hasattr(param, "value") is False:
-                new_params_dict[key] = param
             elif (
                 (key == "x")
                 | (key == "y")
@@ -552,12 +812,12 @@ class ScatterPlots(ObjectDataSelection):
                 | (key == "color")
                 | (key == "size")
             ):
-                if (param.value != "None") & (param.value in self.data_channels.keys()):
-                    new_params_dict[key] = self.data_channels[param.value]
+                if (param != "None") & (param in self.data_channels.keys()):
+                    new_params_dict[key] = self.data_channels[param]
                 else:
                     new_params_dict[key] = None
             else:
-                new_params_dict[key] = param.value
+                new_params_dict[key] = param
 
         ifile = InputFile(
             ui_json=self.params.input_file.ui_json,
@@ -568,43 +828,37 @@ class ScatterPlots(ObjectDataSelection):
         new_params = ScatterPlotParams(input_file=ifile)
 
         driver = ScatterPlotDriver(new_params)
-        generated_figure = go.FigureWidget(driver.run())
+        figure = go.FigureWidget(driver.run())
 
-        self.figure.data = []
-        if len(generated_figure.data) > 0:
-            self.figure.add_trace(generated_figure.data[0])
+        return figure
 
-        self.figure.update_layout(generated_figure.layout)
+    def update_object_options(self, contents):
+        if contents is not None:
+            content_type, content_string = contents.split(",")
+            decoded = io.BytesIO(base64.b64decode(content_string))
+            self.params.geoh5 = Workspace(decoded)
 
-    def update_axes(self, refresh_plot=True):
+        obj_list = self.params.geoh5.objects
 
-        for name in [
-            "x",
-            "y",
-            "z",
-            "color",
-            "size",
-        ]:
-            self.refresh.value = False
+        options = [
+            {"label": obj.parent.name + "/" + obj.name, "value": obj.name}
+            for obj in obj_list
+        ]
+        if len(options) > 0:
+            value = options[0]["value"]
+        else:
+            value = None
 
-            widget = getattr(self, "_" + name)
-            val = widget.value
-            widget.options = list(self.data_channels.keys())
+        return options, value
 
-            if val in list(self.data_channels.keys()):
-                widget.value = val
-            elif "None" in list(self.data_channels.keys()):
-                widget.value = "None"
-            else:
-                widget.value = None
-
-        if refresh_plot:
-            self.refresh.value = True
-
-    def update_choices(self, _):
-        self.refresh.value = False
-
-        obj, _ = self.get_selected_entities()
+    def update_data_options(self, object):
+        obj = None
+        if getattr(
+            self.params, "geoh5", None
+        ) is not None and self.params.geoh5.get_entity(object):
+            for entity in self.params.geoh5.get_entity(object):
+                if isinstance(entity, ObjectBase):
+                    obj = entity
 
         channel_list = ["None"]
         channel_list.extend(obj.get_data_list())
@@ -612,22 +866,38 @@ class ScatterPlots(ObjectDataSelection):
         if "Visual Parameters" in channel_list:
             channel_list.remove("Visual Parameters")
 
+        self.data_channels = {}
         for channel in channel_list:
             self.get_channel(channel)
 
-        self.update_axes(refresh_plot=False)
+        options = list(self.data_channels.keys())
 
-        self.refresh.value = True
-
-    def update_objects(self, _):
-        self.data_channels = {}
-        self.refresh.value = False
-        self.figure.data = []
-        self.refresh.value = True
-
-    def trigger_click(self, _):
-        self.figure.write_html(
-            os.path.join(
-                os.path.abspath(os.path.dirname(self.h5file)), "Crossplot.html"
-            )
+        return (
+            options,
+            options,
+            options,
+            options,
+            options,
+            "None",
+            "None",
+            "None",
+            "None",
+            "None",
         )
+
+    def save_figure(self, fig):
+        buffer = io.StringIO()
+        go.Figure(fig).write_html(buffer)
+        html_bytes = buffer.getvalue().encode()
+        encoded = base64.b64encode(html_bytes).decode()
+        href = "data:text/html;base64," + encoded
+
+        return href
+
+    def run(self):
+        # The reloader has not yet run - open the browser
+        if not environ.get("WERKZEUG_RUN_MAIN"):
+            webbrowser.open_new("http://127.0.0.1:8050/")
+
+        # Otherwise, continue as normal
+        self.app.run_server(host="127.0.0.1", port=8050, debug=False)
